alguma documentação

+++ project_c/includes/reviews.h
@@ -10,7 +10,7 @@ Reviews addReview(Reviews  rev,char * str);
 char * r_getReviewId(Reviews  rev);
 char * r_getUserId(Reviews  rev);
 char * r_getBusinessId(Reviews  rev);
-int r_getStars(Reviews  rev);
+float r_getStars(Reviews  rev);
 int r_getUseful(Reviews  rev);
 int r_getFunny(Reviews  rev);
 int r_getCool(Reviews  rev);
